pyproject.toml
[project]
name = "ayoluchuncloneproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.5.3",
    "daphne>=4.2.1",
    "django>=5.2.4",
    "django-environ>=0.12.0",
    "django-jazzmin>=3.0.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-yasg>=1.21.10",
    "httpx>=0.28.1",
    "pillow>=11.3.0",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
    "redis>=5.2.1",
    "requests>=2.32.4",
]
